pkgmeta: add multiple URL fields
This allows us to specifiy multiple URLs for each project in addition
to the homepage that is defined by the PKGBUILD.

This is similar to what Python packages can do, which then gets
exposed on pypi. But while they allow free form text we specify
a fixed set with a specific purpose, so we can pontentially re-use them
in the future, and also to make it more clear what we expect them to
point to.