base packages: pull the "base" package description from the SRCINFO
When parsing the SRCINFO we only get out the sub packages and the
package fields override the base fields, but the base description
has some vlaue on its own, as it is a description for all contained
packages.

Store the base description away and when fetching the description
for the pacman DB package we look up the description from the SRCINFO
base package if possible.

The onyl downside here is that they come from different sources, so
might be out of sync, but still better then before I guess.