queue: sort by status first
It was sorted by change date before, which does not say when
something was merged but when it was commited.

At least for me, I want to see if there are any problems and what is
currently building when looking at the queue, and only then what already
is there. Might be different for other users.

Sort by the status priority instead. The priority is the same one used to
decide the most important build status to show in the list for the
different build types.